Scope Analysis
This commits adds scoping support to Pythran: a variable can now be
declared locally instead of being declared at the function level. This
is especially useful in presence of OpenMP directives, as it makes them
less verbose.

It's a slightly complex pass, so I documented it in a new
doc/INTERNAL.rst document, validated by the doctest, of course!

Several OpenMP test cases have been updated to reflect this pass's
changes

+++ pythran/pythonic++/core/string.h
@@ -209,7 +209,7 @@ namespace pythonic {
             bool operator==(string_view const & other) const {
                 if(size() != other.size())
                     return false;
-                for(long i=other.get_slice().lower, j=0L;
+                for(size_t i=other.get_slice().lower, j=0L;
                         j<size();
                         i= i + other.get_slice().step, j++)
                     if(other.get_data()[i] != (*this)[j])
